refactoring of highlight list
