+++ src/public/app/layouts/desktop_layout.js
@@ -44,7 +44,7 @@ import BacklinksWidget from "../widgets/floating_buttons/zpetne_odkazy.js";
 import SharedInfoWidget from "../widgets/shared_info.js";
 import FindWidget from "../widgets/find.js";
 import TocWidget from "../widgets/toc.js";
-import HighlightedTextWidget from "../widgets/highlighted_text.js";
+import HighlightsListWidget from "../widgets/highlights_list.js";
 import BulkActionsDialog from "../widgets/dialogs/bulk_actions.js";
 import AboutDialog from "../widgets/dialogs/about.js";
 import HelpDialog from "../widgets/dialogs/help.js";
@@ -185,7 +185,7 @@ export default class DesktopLayout {
                     )
                     .child(new RightPaneContainer()
                         .child(new TocWidget())
-                        .child(new HighlightedTextWidget())
+                        .child(new HighlightsListWidget())
                         .child(...this.customWidgets.get('right-pane'))
                     )
                 )
